--- src/styles/styles.js
@@ -1,274 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-export const styles = StyleSheet.create({
-    container: {
-        flex: 1,
-        backgroundColor: '#aaaaaa',
-        paddingTop: 50,
-        paddingHorizontal: 20,
-    },
-    title: {
-        fontSize: 30,
-        fontWeight: 'bold',
-        textAlign: 'center',
-        marginBottom: 15,
-        marginTop: 15,
-        color: '#fff',
-    },
-    inputContainer: {
-        flexDirection: 'row',
-        marginTop: 10,
-        marginBottom: 20,
-    },
-    input: {
-        flex: 1,
-        borderWidth: 1,
-        borderColor: '#ddd',
-        borderRadius: 8,
-        padding: 14,
-        backgroundColor: '#fff',
-        marginRight: 10,
-        fontSize: 18,
-    },
-    addButton: {
-        backgroundColor: '#007AFF',
-        paddingHorizontal: 20,
-        paddingVertical: 12,
-        borderRadius: 8,
-        justifyContent: 'center',
-    },
-    addButtonDisabled: {
-        backgroundColor: '#ccc',
-    },
-    addButtonText: {
-        color: '#fff',
-        fontWeight: 'bold',
-        fontSize: 16,
-    },
-    taskList: {
-        flex: 1,
-    },
-    taskListContent: {
-        paddingBottom: 20,
-    },
-    taskContainer: {
-        marginBottom: 5,
-    },
-    parentTaskItem: {
-        flexDirection: 'row',
-        alignItems: 'center',
-        backgroundColor: '#fff',
-        paddingHorizontal: 15,
-        paddingVertical: 10,
-        borderRadius: 8,
-        shadowColor: '#000',
-        shadowOffset: { width: 0, height: 1 },
-        shadowOpacity: 0.1,
-        shadowRadius: 2,
-        elevation: 2,
-        borderLeftWidth: 6,
-        borderLeftColor: '#007AFF',
-    },
-    parentTaskItemSelected: {
-        backgroundColor: '#E8F5E8',
-        borderWidth: 2,
-        borderColor: '#4CAF50',
-    },
-    childTaskContainer: {
-        marginLeft: 25,
-        marginTop: 5,
-    },
-    childTaskItem: {
-        flexDirection: 'row',
-        alignItems: 'center',
-        backgroundColor: '#f0f0f0',
-        paddingHorizontal: 15,
-        paddingVertical: 8,
-        borderRadius: 6,
-        borderLeftWidth: 6,
-        borderLeftColor: '#4CAF50',
-    },
-    checkboxContainer: {
-        padding: 8,
-    },
-    checkbox: {
-        width: 28,
-        height: 28,
-        borderWidth: 2,
-        borderColor: '#ddd',
-        borderRadius: 4,
-        justifyContent: 'center',
-        alignItems: 'center',
-    },
-    checkboxCompleted: {
-        backgroundColor: '#007AFF',
-        borderColor: '#007AFF',
-    },
-    checkboxText: {
-        color: '#fff',
-        fontWeight: 'bold',
-        fontSize: 16,
-    },
-    taskTextContainer: {
-        flex: 1,
-        paddingVertical: 8,
-        paddingHorizontal: 12,
-    },
-    taskTextTouchable: {
-        flex: 1,
-        paddingVertical: 4,
-    },
-    taskText: {
-        fontSize: 16,
-        color: '#333',
-    },
-    taskTextCompleted: {
-        textDecorationLine: 'line-through',
-        color: '#999',
-    },
-    editContainer: {
-        flex: 1,
-    },
-    editInput: {
-        borderWidth: 1,
-        borderColor: '#2196F3',
-        borderRadius: 4,
-        padding: 8,
-        backgroundColor: '#fff',
-        fontSize: 16,
-        marginBottom: 8,
-    },
-    editButtons: {
-        flexDirection: 'row',
-        justifyContent: 'flex-end',
-    },
-    saveButton: {
-        backgroundColor: '#4CAF50',
-        paddingHorizontal: 12,
-        paddingVertical: 9,
-        borderRadius: 4,
-        marginRight: 8,
-    },
-    saveButtonText: {
-        color: '#fff',
-        fontSize: 14,
-        fontWeight: 'bold',
-    },
-    cancelEditButton: {
-        backgroundColor: '#FF5722',
-        paddingHorizontal: 12,
-        paddingVertical: 9,
-        borderRadius: 4,
-        marginRight: 8,
-    },
-    cancelEditButtonText: {
-        color: '#fff',
-        fontSize: 14,
-        fontWeight: 'bold',
-    },
-    deleteButton: {
-        backgroundColor: '#FF3B30',
-        paddingHorizontal: 12,
-        paddingVertical: 9,
-        borderRadius: 4,
-    },
-    deleteButtonText: {
-        color: '#fff',
-        fontSize: 14,
-        fontWeight: 'bold',
-    },
-    childAddButtonInline: {
-        backgroundColor: '#4CAF50',
-        paddingHorizontal: 12,
-        paddingVertical: 6,
-        borderRadius: 4,
-        justifyContent: 'center',
-        alignItems: 'center',
-        marginLeft: 8,
-        marginRight: 8,
-        minWidth: 32,
-        height: 32,
-    },
-    childAddButtonInlineSelected: {
-        backgroundColor: '#2196F3',
-    },
-    childAddButtonInlineText: {
-        color: '#fff',
-        fontSize: 14,
-        fontWeight: 'bold',
-    },
-    editButton: {
-        backgroundColor: '#FF9800',
-        paddingHorizontal: 12,
-        paddingVertical: 9,
-        borderRadius: 4,
-        marginLeft: 4,
-    },
-    editButtonText: {
-        color: '#fff',
-        fontSize: 14,
-        fontWeight: 'bold',
-    },
-    draggedItem: {
-        opacity: 0.5,
-        backgroundColor: '#e0e0e0',
-    },
-    dragIndicator: {
-        position: 'absolute',
-        right: 5,
-        top: 5,
-    },
-    dragIndicatorText: {
-        fontSize: 16,
-    },
-    dropZoneBetween: {
-        backgroundColor: '#F0F8FF',
-        paddingVertical: 8,
-        paddingHorizontal: 16,
-        marginVertical: 4,
-        borderRadius: 4,
-        borderWidth: 1,
-        borderColor: '#2196F3',
-        borderStyle: 'dashed',
-        alignItems: 'center',
-    },
-    dropZoneFinal: {
-        backgroundColor: '#F0F8FF',
-        paddingVertical: 12,
-        paddingHorizontal: 16,
-        marginTop: 8,
-        borderRadius: 4,
-        borderWidth: 1,
-        borderColor: '#2196F3',
-        borderStyle: 'dashed',
-        alignItems: 'center',
-    },
-    dropZoneLabel: {
-        color: '#2196F3',
-        fontSize: 12,
-        fontWeight: 'bold',
-    },
-    childDropZone: {
-        backgroundColor: '#F0F8FF',
-        paddingHorizontal: 8,
-        paddingVertical: 4,
-        marginHorizontal: 4,
-        borderRadius: 4,
-        borderWidth: 1,
-        borderColor: '#2196F3',
-        borderStyle: 'dashed',
-        alignItems: 'center',
-        justifyContent: 'center',
-    },
-    childDropZoneText: {
-        color: '#2196F3',
-        fontSize: 10,
-        fontWeight: 'bold',
-    },
-    counter: {
-        textAlign: 'center',
-        marginBottom: 20,
-        fontSize: 16,
-        color: '#666',
-    },
-});
